Bug fixes; move time_from_name to be topic specific
